Enable PY_SSIZE_T_CLEAN in order to avoid deprecation warning and/or segfault under Python 3.8.0b1 (https://github.com/oracle/python-cx_Oracle/issues/317).

+++ src/cxoConnection.c
@@ -1199,7 +1199,7 @@ static PyObject *cxoConnection_commit(cxoConnection *conn, PyObject *args)
 //-----------------------------------------------------------------------------
 static PyObject *cxoConnection_begin(cxoConnection *conn, PyObject *args)
 {
-    uint32_t transactionIdLength, branchIdLength;
+    Py_ssize_t transactionIdLength, branchIdLength;
     const char *transactionId, *branchId;
     int formatId, status;
